% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{get_nhood_size}
\alias{get_nhood_size}
\title{Calculate size of neighborhood}
\usage{
get_nhood_size(fwhm = NULL, vDims = NULL, brainMask, verbose = TRUE)
}
\arguments{
\item{fwhm}{numeric value of full width at half maximum}

\item{vDims}{numeric vector of voxel dimensions}

\item{brainMask}{antsImage}

\item{verbose}{TRUE or FALSE}
}
\value{
list of voxel properties
}
\description{
Calculate size of neighborhood
}
\examples{
\dontrun{
get_nhood_size(fwhm = 3, vDims = c(2, 2, 2),
brainMask = "inst/extdata/gm10_pcals_rest.nii.gz", verbose = TRUE)
}
}
