% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{get_pvals_by_voxel}
\alias{get_pvals_by_voxel}
\title{Calculates pvals for each voxel}
\usage{
get_pvals_by_voxel(voxel_vector, predictors)
}
\arguments{
\item{voxel_vector}{vector of length 1xn, where n is number of subjects}

\item{predictors}{design matrix of predictors}
}
\value{
vector of p-values from various models
}
\description{
Calculates pvals for each voxel
}
\examples{
\dontrun{
get_pvals_by_voxel(1:5, matrix(1:10, nrow = 5))
}
}
