% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{get_weights}
\alias{get_weights}
\title{Assign weights to each voxel in neighborhood}
\usage{
get_weights(offsets, voxelDims, sigma)
}
\arguments{
\item{offsets}{numerical vector of voxel distances from central voxel}

\item{voxelDims}{numerical vector of voxel dimensions}

\item{sigma}{numeric defined from fwhm}
}
\value{
numerical vector of weights defined based on sigma
}
\description{
Assign weights to each voxel in neighborhood for weighted PCA/regression.
Weights are assigned according to sigma (generated from FWHM)
}
\examples{
\dontrun{
get_weights(offsets = c(3, 1, 3), voxelDims = c(2, 2, 2), sigma = sigma)
}
}
