% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco}
\alias{imco}
\title{Performs Intermodal Coupling}
\usage{
imco(
  files,
  brainMask,
  out_dir,
  out_name,
  subMask = NULL,
  ref = 1,
  fwhm = NULL,
  thresh = 0.005,
  radius = NULL,
  reverse = FALSE,
  verbose = TRUE,
  outDir = NULL,
  propMiss = NULL,
  corstr = NULL,
  pcaType = NULL,
  matrixType = NULL,
  returnNeighborhoods = FALSE
)
}
\arguments{
\item{files}{list of lists, A_i, where A_i is a list of file paths for modality i}

\item{brainMask}{antsImage mask to apply before coupling (optional)}

\item{out_dir}{string for output directory}

\item{out_name}{TODO}

\item{subMask}{TODO}

\item{ref}{TODO}

\item{fwhm}{numerical value of full width at half maximum for calculating neighborhood size}

\item{thresh}{TODO}

\item{radius}{TODO}

\item{reverse}{TODO}

\item{verbose}{TRUE or FALSE}

\item{outDir}{TODO}

\item{propMiss}{TODO}

\item{corstr}{TODO}

\item{pcaType}{"global" or "unscaled"}

\item{matrixType}{"wcov" (weighted covariance) or "wcor" (weighted correlation)}

\item{returnNeighborhoods}{TODO}
}
\value{
list of coupled antsImages and file paths
}
\description{
Performs Intermodal Coupling
}
\examples{
\dontrun{
mod1_files <- list("mod1_subj1.nii.gz", "mod1_subj2.nii.gz")
mod2_files <- list("mod2_subj1.nii.gz", "mod2_subj2.nii.gz")
mod3_files <- list("mod3_subj1.nii.gz", "mod3_subj2.nii.gz")
imco(files = list(mod1_files, mod2_files, mod3_files),
    brainMask = "inst/extdata/gm10_pcals_rest.nii.gz",
    out_dir = coupled_images, out_name = "", type = "pca", fwhm = 3)
}
}
