% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco_pca}
\alias{imco_pca}
\title{Intermodal Coupling with PCA}
\usage{
imco_pca(
  files,
  nhoods,
  nhood_weights,
  mask_indices,
  verbose = TRUE,
  full_pca_dir = NULL,
  prop_miss = NULL,
  pca_type = NULL,
  matrix_type = NULL,
  return_neighborhoods = FALSE
)
}
\arguments{
\item{files}{TODO}

\item{nhoods}{TODO}

\item{nhood_weights}{TODO}

\item{mask_indices}{TODO}

\item{verbose}{TRUE or FALSE}

\item{full_pca_dir}{TODO}

\item{prop_miss}{TODO}

\item{pca_type}{"global" or "unscaled"}

\item{matrix_type}{"wcov" or "wcor"}

\item{return_neighborhoods}{TRUE or FALSE}
}
\value{
TODO
}
\description{
Intermodal Coupling with PCA
}
\examples{
\dontrun{
TODO
}
}
