% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{make_ants_image}
\alias{make_ants_image}
\title{Makes ANTs Image in mask out of vector of intensities}
\usage{
make_ants_image(vec, mask_indices, reference)
}
\arguments{
\item{vec}{numerical vector of intensities}

\item{mask_indices}{binary array - 1 in mask and 0 out of mask}

\item{reference}{antsImage to supply metadata}
}
\value{
antsImage
}
\description{
Makes ANTs Image in mask out of vector of intensities
}
\examples{
\dontrun{
TODO
}
}
