% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{write_pvals}
\alias{write_pvals}
\title{Writes p-values to antsImage}
\usage{
write_pvals(image_list, mask, dir, is_descriptive = FALSE, out_dir)
}
\arguments{
\item{image_list}{TODO}

\item{mask}{TODO}

\item{dir}{TODO}

\item{is_descriptive}{TODO}

\item{out_dir}{TODO}
}
\value{
TODO
}
\description{
Writes p-values to antsImage
}
\examples{
\dontrun{
TODO
}
}
