% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco}
\alias{imco}
\title{Performs PCA-based Intermodal Coupling}
\usage{
imco(
  files,
  brain_mask,
  out_dir = NULL,
  out_name = NULL,
  fwhm = 3,
  verbose = TRUE,
  prop_miss = 0.9,
  use_ratio = FALSE,
  return_coupling = FALSE
)
}
\arguments{
\item{files}{list of paths to niftis (string) or list of antsImages}

\item{brain_mask}{path to mask nifti or antsImage mask to apply before coupling (optional)}

\item{out_dir}{string path to output directory}

\item{out_name}{string prefix for file name. Final name is "'out_name'_coupling.nii.gz"}

\item{fwhm}{numerical value of full width at half maximum for calculating neighborhood size; default is 3}

\item{verbose}{TRUE or FALSE; default TRUE}

\item{prop_miss}{proportion of voxels in a neighborhood allowed to be missing; default is 0.9}

\item{use_ratio}{default FALSE for (logit of 1st eigenvalue scaled to 0-1), TRUE for (1st eigenvalue/total variance)}

\item{return_coupling}{default FALSE to write image to out_dir, TRUE to return coupling image without writing}
}
\value{
list of coupled antsImages and file paths
}
\description{
Performs PCA-based Intermodal Coupling
}
\examples{
\dontrun{
mod1_filepath <- "/path/to/proj/dir/mod1_subj1.nii.gz"
mod2_filepath <- "/path/to/proj/dir/mod2_subj1.nii.gz"
mod3_filepath <- "/path/to/proj/dir/mod3_subj1.nii.gz"
imco(files = list(mod1_files, mod2_files, mod3_files),
    brain_mask = "/path/to/proj/dir/grey_matter_mask.nii.gz",
    out_dir = "/path/to/proj/dir/coupled_images",
    out_name = "subj-1",
    fwhm = 3)
}
}
