% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco}
\alias{imco}
\title{Performs Intermodal Coupling}
\usage{
imco(
  files,
  brain_mask,
  out_dir = NULL,
  out_name = NULL,
  fwhm = NULL,
  verbose = TRUE,
  prop_miss = 1,
  pca_type = c("global_wcov", "unscaled_wcor"),
  use_ratio = FALSE,
  return_coupling = FALSE
)
}
\arguments{
\item{files}{list of string paths to files to couple}

\item{brain_mask}{antsImage mask to apply before coupling (optional)}

\item{out_dir}{string path to output directory}

\item{out_name}{string prefix for file name. Final name is "outname_coupling.nii.gz"}

\item{fwhm}{numerical value of full width at half maximum for calculating neighborhood size}

\item{verbose}{TRUE or FALSE}

\item{prop_miss}{proportion of voxels in a neighborhood allowed to be missing}

\item{pca_type}{"global_wcov" or "unscaled_wcor"}

\item{use_ratio}{default FALSE for (logit of 1st eigenvalue scaled to 0-1), TRUE for (1st eigenvalue/total variance)}

\item{return_coupling}{default FALSE to write image to out_dir, TRUE to return coupling image without writing}
}
\value{
list of coupled antsImages and file paths
}
\description{
Performs Intermodal Coupling
}
\examples{
\dontrun{
mod1_files <- list("mod1_subj1.nii.gz")
mod2_files <- list("mod2_subj1.nii.gz")
mod3_files <- list("mod3_subj1.nii.gz")
imco(files = list(mod1_files, mod2_files, mod3_files),
    brainMask = "inst/extdata/gm10_pcals_rest.nii.gz",
    out_dir = coupled_images, out_name = "", type = "pca", fwhm = 3)
}
}
