% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco_pca.R
\name{imco_pca}
\alias{imco_pca}
\title{Intermodal Coupling with PCA}
\usage{
imco_pca(
  files,
  nhoods,
  nhood_weights,
  mask_indices,
  verbose = TRUE,
  prop_miss = NULL,
  pca_type = NULL,
  use_ratio
)
}
\arguments{
\item{files}{list of antsImages}

\item{nhoods}{from antsGetNeighborhood in mask}

\item{nhood_weights}{from get_weights}

\item{mask_indices}{indices at which mask value is 1}

\item{verbose}{TRUE or FALSE}

\item{prop_miss}{proportion of voxels in a neighborhood allowed to be missing}

\item{pca_type}{"global_wcov" or "unscaled_wcor"}

\item{use_ratio}{default FALSE for (logit of 1st eigenvalue scaled to 0-1), TRUE for (1st eigenvalue/total variance)}
}
\value{
antsImage with couplinh values
}
\description{
Intermodal Coupling with PCA
}
