% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco_pca}
\alias{imco_pca}
\title{Intermodal Coupling with PCA}
\usage{
imco_pca(
  files,
  nhoods,
  nWts,
  mask_indices,
  verbose = TRUE,
  full_pca_dir = NULL,
  propMiss = NULL,
  pcaType = NULL,
  matrixType = NULL,
  returnNeighborhoods = FALSE
)
}
\arguments{
\item{files}{TODO}

\item{nhoods}{TODO}

\item{nWts}{TODO}

\item{mask_indices}{TODO}

\item{verbose}{TRUE or FALSE}

\item{full_pca_dir}{TODO}

\item{propMiss}{TODO}

\item{pcaType}{"global" or "unscaled"}

\item{matrixType}{"wcov" or "wcor"}

\item{returnNeighborhoods}{TRUE or FALSE}
}
\value{
TODO
}
\description{
Intermodal Coupling with PCA
}
\examples{
\dontrun{
TODO
}
}
