% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ants_image.R
\name{make_ants_image}
\alias{make_ants_image}
\title{Makes ANTs Image in mask out of vector of intensities}
\usage{
make_ants_image(vec, mask_indices, reference)
}
\arguments{
\item{vec}{numerical vector of intensities with 0s removed}

\item{mask_indices}{vector of indices at which mask is 1}

\item{reference}{antsImage to supply metadata}
}
\value{
antsImage
}
\description{
Makes ANTs Image in mask out of vector of intensities
}
\examples{
\dontrun{
TODO
}
}
\keyword{internal}
